Move deserialize_tools_inplace back to original import path
